.history/app/requirements_20250704101135.txt
# requirements.txt

# Core library untuk Deep Learning
tensorflow==2.16.1

# Library untuk komputasi numerik, seringkali dependency dari TensorFlow
numpy==1.26.4

# Untuk memanipulasi dan membuka gambar
Pillow==10.3.0

# Untuk membuat plot dan visualisasi data
matplotlib==3.9.0
seaborn==0.13.2

# Untuk evaluasi model (confusion matrix, classification report)
scikit-learn==1.5.0

# Untuk membuat aplikasi web interaktif
streamlit==1.36.0